Fix type casting in resourceId calculation
warning CS0675: Bitwise-or operator used on a sign-extended operand; consider casting to a smaller unsigned type first

+++ src/ApkReader/ArscReader.cs
@@ -84,7 +84,7 @@ public static class ArscReader
                                 {
                                     continue;
                                 }
-                                var resourceId = (packageHeader.Id << 24) | (table.RawID << 16) | i;
+                                var resourceId = (packageHeader.Id << 24) | (table.RawID << 16) | (uint)i;
                                 var entry = reader.ReadResTable_entry();
                                 entries[i] = entry;
                                 if ((entry.Flags & EntryFlags.FLAG_COMPLEX) == 0)
